requirements.txt
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
build==1.3.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.0
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
durationpy==0.10
easyocr==1.7.2
filelock==3.20.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
google-auth==2.45.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.2.0
langchain-core==1.2.6
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
lazy_loader==0.4
llama_cpp_python==0.3.16
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
posthog==5.4.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pyclipper==1.4.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.26.7
pyparsing==3.3.1
pypdfium2==5.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tabula-py==2.10.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2026.1.14
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typer==0.21.0
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
